Move Asset conditions
