+++ src/Asset/Conditions/FilenameConditionRule.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace craft\elements\conditions\assets;
+namespace CraftCms\Cms\Asset\Conditions;
 
-use craft\base\conditions\BaseNumberConditionRule;
+use craft\base\conditions\BaseTextConditionRule;
 use craft\base\ElementInterface;
 use craft\elements\conditions\ElementConditionRuleInterface;
 use craft\elements\db\AssetQuery;
@@ -11,19 +11,19 @@ use CraftCms\Cms\Element\Queries\Contracts\ElementQueryInterface;
 use function CraftCms\Cms\t;
 
 /**
- * Width condition rule.
+ * Filename condition rule.
  *
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 4.0.0
  */
-class WidthConditionRule extends BaseNumberConditionRule implements ElementConditionRuleInterface
+class FilenameConditionRule extends BaseTextConditionRule implements ElementConditionRuleInterface
 {
     /**
      * @inheritdoc
      */
     public function getLabel(): string
     {
-        return t('Width');
+        return t('Filename');
     }
 
     /**
@@ -31,7 +31,7 @@ class WidthConditionRule extends BaseNumberConditionRule implements ElementCondi
      */
     public function getExclusiveQueryParams(): array
     {
-        return ['width'];
+        return ['filename'];
     }
 
     /**
@@ -40,7 +40,7 @@ class WidthConditionRule extends BaseNumberConditionRule implements ElementCondi
     public function modifyQuery(ElementQueryInterface $query): void
     {
         /** @var AssetQuery $query */
-        $query->width($this->paramValue());
+        $query->filename($this->paramValue());
     }
 
     /**
@@ -49,6 +49,6 @@ class WidthConditionRule extends BaseNumberConditionRule implements ElementCondi
     public function matchElement(ElementInterface $element): bool
     {
         /** @var Asset $element */
-        return $this->matchValue($element->getWidth());
+        return $this->matchValue($element->getFilename());
     }
 }
