+++ src/Asset/Conditions/HeightConditionRule.php
@@ -1,31 +1,29 @@
 <?php
 
-namespace craft\elements\conditions\assets;
+namespace CraftCms\Cms\Asset\Conditions;
 
-use Craft;
-use craft\base\conditions\BaseMultiSelectConditionRule;
+use craft\base\conditions\BaseNumberConditionRule;
 use craft\base\ElementInterface;
 use craft\elements\conditions\ElementConditionRuleInterface;
 use craft\elements\db\AssetQuery;
 use CraftCms\Cms\Asset\Elements\Asset;
 use CraftCms\Cms\Element\Queries\Contracts\ElementQueryInterface;
-use CraftCms\Cms\Support\Arr;
 use function CraftCms\Cms\t;
 
 /**
- * Asset volume condition rule.
+ * Height condition rule.
  *
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 4.0.0
  */
-class VolumeConditionRule extends BaseMultiSelectConditionRule implements ElementConditionRuleInterface
+class HeightConditionRule extends BaseNumberConditionRule implements ElementConditionRuleInterface
 {
     /**
      * @inheritdoc
      */
     public function getLabel(): string
     {
-        return t('Volume');
+        return t('Height');
     }
 
     /**
@@ -33,16 +31,7 @@ class VolumeConditionRule extends BaseMultiSelectConditionRule implements Elemen
      */
     public function getExclusiveQueryParams(): array
     {
-        return ['volume', 'volumeId'];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function options(): array
-    {
-        $volumes = Craft::$app->getVolumes()->getAllVolumes();
-        return Arr::pluck($volumes, 'name', 'uid');
+        return ['height'];
     }
 
     /**
@@ -51,8 +40,7 @@ class VolumeConditionRule extends BaseMultiSelectConditionRule implements Elemen
     public function modifyQuery(ElementQueryInterface $query): void
     {
         /** @var AssetQuery $query */
-        $volumes = Craft::$app->getVolumes();
-        $query->volumeId($this->paramValue(fn($uid) => $volumes->getVolumeByUid($uid)->id ?? null));
+        $query->height($this->paramValue());
     }
 
     /**
@@ -61,6 +49,6 @@ class VolumeConditionRule extends BaseMultiSelectConditionRule implements Elemen
     public function matchElement(ElementInterface $element): bool
     {
         /** @var Asset $element */
-        return $this->matchValue($element->getVolume()->uid);
+        return $this->matchValue($element->getHeight());
     }
 }
