+++ src/Asset/Conditions/WidthConditionRule.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace craft\elements\conditions\assets;
+namespace CraftCms\Cms\Asset\Conditions;
 
-use craft\base\conditions\BaseDateRangeConditionRule;
+use craft\base\conditions\BaseNumberConditionRule;
 use craft\base\ElementInterface;
 use craft\elements\conditions\ElementConditionRuleInterface;
 use craft\elements\db\AssetQuery;
@@ -11,19 +11,19 @@ use CraftCms\Cms\Element\Queries\Contracts\ElementQueryInterface;
 use function CraftCms\Cms\t;
 
 /**
- * Date Modified condition rule.
+ * Width condition rule.
  *
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 4.0.0
  */
-class DateModifiedConditionRule extends BaseDateRangeConditionRule implements ElementConditionRuleInterface
+class WidthConditionRule extends BaseNumberConditionRule implements ElementConditionRuleInterface
 {
     /**
      * @inheritdoc
      */
     public function getLabel(): string
     {
-        return t('File Modification Date');
+        return t('Width');
     }
 
     /**
@@ -31,7 +31,7 @@ class DateModifiedConditionRule extends BaseDateRangeConditionRule implements El
      */
     public function getExclusiveQueryParams(): array
     {
-        return ['dateModified'];
+        return ['width'];
     }
 
     /**
@@ -40,7 +40,7 @@ class DateModifiedConditionRule extends BaseDateRangeConditionRule implements El
     public function modifyQuery(ElementQueryInterface $query): void
     {
         /** @var AssetQuery $query */
-        $query->dateModified($this->queryParamValue());
+        $query->width($this->paramValue());
     }
 
     /**
@@ -49,6 +49,6 @@ class DateModifiedConditionRule extends BaseDateRangeConditionRule implements El
     public function matchElement(ElementInterface $element): bool
     {
         /** @var Asset $element */
-        return $this->matchValue($element->dateModified);
+        return $this->matchValue($element->getWidth());
     }
 }
